# server/integrations/apiConnector/openapi.yaml
openapi: 3.0.3
info:
  title: Lead2Lease API Connector
  description: |
    REST API for external PMS and integrators to sync leads, conversations, tours, and listings.

    ## Setup
    1. Go to **Integrations** → **API** → **Configure**
    2. Click **Generate Key** (optionally enter a name like "Production")
    3. Copy the key immediately — it is shown only once
    4. Use `Authorization: Bearer <your_api_key>` in all requests
  version: 1.0.0

servers:
  - url: http://localhost:5000/api/integrations/api/v1
    description: Local development
  - url: https://app.lead2lease.ai/api/integrations/api/v1
    description: Production
  - url: /api/integrations/api/v1
    description: Relative path (use your instance origin)

security:
  - BearerAuth: []

paths:
  /leads:
    get:
      summary: List leads
      parameters:
        - name: updated_since
          in: query
          schema:
            type: string
            format: date-time
        - name: source
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  meta:
                    type: object
                    properties:
                      requestId:
                        type: string
                      nextCursor:
                        type: string
                      count:
                        type: integer

  /leads/{leadId}:
    get:
      summary: Get lead by ID
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lead details
        '404':
          description: Lead not found

  /conversations:
    get:
      summary: List conversations for a lead
      parameters:
        - name: leadId
          in: query
          required: true
          schema:
            type: string
        - name: cursor
          in: query
        - name: limit
          in: query
      responses:
        '200':
          description: List of conversations

  /conversations/{conversationId}/messages:
    get:
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of messages
        '404':
          description: Conversation not found

  /tours/bookings:
    get:
      summary: List tour bookings
      parameters:
        - name: leadId
          in: query
        - name: propertyId
          in: query
      responses:
        '200':
          description: List of bookings

  /properties:
    get:
      summary: List properties
      responses:
        '200':
          description: List of properties

  /properties/{propertyId}:
    get:
      summary: Get property by ID
      parameters:
        - name: propertyId
          in: path
          required: true
      responses:
        '200':
          description: Property details
        '404':
          description: Property not found

  /properties/{propertyId}/units:
    get:
      summary: List units for a property
      parameters:
        - name: propertyId
          in: path
          required: true
      responses:
        '200':
          description: List of units

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: Use your API key as the Bearer token. Create keys in Integrations > API.

  schemas:
    Lead:
      type: object
      properties:
        id:
          type: string
        externalId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        source:
          type: object
          properties:
            channel:
              type: string
        propertyId:
          type: string
        unitId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time